Add same logic to processors.
