+++ libbeat/processors/actions/append.go
@@ -24,6 +24,7 @@ import (
 	"github.com/elastic/beats/v7/libbeat/processors"
 	"github.com/elastic/beats/v7/libbeat/processors/checks"
 	jsprocessor "github.com/elastic/beats/v7/libbeat/processors/script/javascript/module/processor"
+	"github.com/elastic/beats/v7/libbeat/publisher"
 	conf "github.com/elastic/elastic-agent-libs/config"
 	"github.com/elastic/elastic-agent-libs/logp"
 )
@@ -81,7 +82,9 @@ func (f *appendProcessor) Run(event *beat.Event) (*beat.Event, error) {
 	err := f.appendValues(f.config.TargetField, f.config.Fields, f.config.Values, event)
 	if err != nil {
 		errMsg := fmt.Errorf("failed to append fields in append processor: %w", err)
-		f.logger.Debug(errMsg.Error())
+		if publisher.LogWithTrace() {
+			f.logger.Debug(errMsg.Error())
+		}
 		if f.config.FailOnError {
 			event = backup
 			if _, err := event.PutValue("error.message", errMsg.Error()); err != nil {
