+++ libbeat/processors/actions/decode_base64_field.go
@@ -28,6 +28,7 @@ import (
 	"github.com/elastic/beats/v7/libbeat/processors"
 	"github.com/elastic/beats/v7/libbeat/processors/checks"
 	jsprocessor "github.com/elastic/beats/v7/libbeat/processors/script/javascript/module/processor"
+	"github.com/elastic/beats/v7/libbeat/publisher"
 	cfg "github.com/elastic/elastic-agent-libs/config"
 	"github.com/elastic/elastic-agent-libs/logp"
 	"github.com/elastic/elastic-agent-libs/mapstr"
@@ -84,7 +85,9 @@ func (f *decodeBase64Field) Run(event *beat.Event) (*beat.Event, error) {
 	err := f.decodeField(event)
 	if err != nil {
 		errMsg := fmt.Errorf("failed to decode base64 fields in processor: %v", err)
-		f.log.Debug(errMsg.Error())
+		if publisher.LogWithTrace() {
+			f.log.Debug(errMsg.Error())
+		}
 		if f.config.FailOnError {
 			event = backup
 			event.PutValue("error.message", errMsg.Error())
